requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cloudinary==1.29.0
dj-database-url==1.0.0
Django==4.1
django-ckeditor==6.5.1
django-comments-xtd==2.9.7
django-contrib-comments==2.2.0
django-crispy-forms==1.14.0
django-etc==1.3.1
django-hitcount==1.3.5
django-js-asset==2.0.0
django-on-heroku==1.1.2
django-social-share==2.3.0
djangorestframework==3.13.1
docutils==0.19
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-magic-bin==0.4.14
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
